modular redux based on DUCK principles
