+++ .eslintrc
@@ -11,11 +11,14 @@
     "es6": true,
     "classes": true
   },
+  "globals": {
+    "browser": true
+  },
   "rules": {
     "semi": [2, "never"],
     "import/no-extraneous-dependencies": [2, { "devDependencies": true }],
+    "jsx-a11y/no-static-element-interactions": [0],
     "arrow-body-style": [0],
-    "react/prefer-stateless-function": [1],
     "react/no-unknown-property": [0],
     "react/react-in-jsx-scope": [0],
     "react/jsx-no-bind": [0],
@@ -23,7 +26,6 @@
     "react/self-closing-comp": [0],
     "react/prop-types": [0],
     "generator-star-spacing": [0],
-    "jsx-a11y/no-static-element-interactions": [1],
-    "no-unused-vars": [0]
+    "no-unused-vars": [1],
   }
 }
