+++ README.md
@@ -4,6 +4,6 @@ Modified **create-react-app** with following packages:
 
 - mocha
 - chai
-- immutable (optional now with *redux*)
-- redux (TODO: create basic implementation)
-- react-router (basic routing)
+- immutable
+- redux
+- react-router
