--- src/actions/actions.js
@@ -1,7 +1,0 @@
-export const DEFAULT_ACTION = 'DEFAULT_ACTION'
-
-export const defaultAction = () => {
-  return {
-    type: DEFAULT_ACTION,
-  }
-}
